requirements.txt
python-dotenv==0.19.2
version_parser==1.0.1
hurry.filesize==0.9
scandir==1.7
grpcio-tools==1.34.1
# py3exiv2==0.9.3
packaging==21.3
version-parser==1.0.1
# supervisely==6.4.3
